function/requirements.txt
Pillow==9.0.1
numpy==1.22.3
aws-lambda-typing==2.15.0